# docs/swagger.yaml
definitions:
...
  models.Record:
    properties:
      create_time:
        example: "2020-09-27"
        type: string
      id:
        example: 1
        type: integer
      language:
        example: English
        type: string
      problem:
        example: 1
        type: integer
      sucess:
        example: true
        type: boolean
      username:
        example: Peter
        type: string
    type: object
  models.User:
    properties:
      id:
        example: 1
        type: integer
      password:
        example: xxxxxxxxx
        type: string
      usernmae:
        example: Peter
        type: string
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: User login function
      parameters:
      - description: User login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: User login
      tags:
      - users
...
  /record:
    post:
      consumes:
      - application/json
      description: Create a new record with the input paylod
      parameters:
      - description: Create record
        in: body
        name: record
        required: true
        schema:
          $ref: '#/definitions/models.Record'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Create a new record
      tags:
      - records
...
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Register a new user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Register a new user
      tags:
      - users
...
swagger: "2.0"